+++ analysisapi/.gitignore
@@ -1,6 +1,3 @@
-__pycache__/
-venv/
-chroma_db/
-
-# 一時的に作成しているファイルを無視
-tmp/
+__pycache__
+venv
+chroma_db
