+++ analysisapi/__main__.py
@@ -1,0 +1,6 @@
+"""Analysis API パッケージを呼び出した際に処理されるモジュールです。"""
+
+from .app import main
+
+if __name__ == "__main__":
+    main()
