+++ analysisapi/applicationcore/client/client_constants.py
@@ -1,0 +1,14 @@
+"""クライアントの定数を定義するクラスです。"""
+
+
+class ClientConstants:
+    """
+    クライアントの定数を定義するクラスです。
+    """
+
+    WEATHER_DATA_FETCH_TIME_URL = (
+        "https://www.jma.go.jp/bosai/amedas/data/latest_time.txt"
+    )
+    AMEDAS_STATION_LOCATION_URL = (
+        "https://www.jma.go.jp/bosai/amedas/const/amedastable.json"
+    )
