+++ analysisapi/generateopenapi/__main__.py
@@ -1,3 +1,6 @@
-if __name__ == '__main__':
-    from .cli import main
+"""generateopenapiのモジュールです。"""
+
+from .generate_openapi import main
+
+if __name__ == "__main__":
     main()
