+++ analysisapi/generateopenapi/generate_openapi.py
@@ -1,0 +1,23 @@
+"""OpenAPI 仕様書を生成するモジュールです。"""
+
+import json
+import os
+from fastapi import FastAPI
+import settings
+from web.controller import Controller
+
+
+def main():
+    """
+    OpenAPI仕様書を生成します。
+    """
+    app = FastAPI(
+        title=settings.PROJECT_NAME,
+        description=settings.PROJECT_DESCRIPTION,
+        version=settings.PROJECT_VERSION,
+    )
+    app.include_router(Controller().router, prefix="/api", tags=["analysisapi"])
+    os.makedirs(os.path.dirname(settings.API_SPECIFICATION_FILE_PATH), exist_ok=True)
+    with open(settings.API_SPECIFICATION_FILE_PATH, "w", encoding="utf-8") as f:
+        api_spec = app.openapi()
+        f.write(json.dumps(api_spec, indent=2, ensure_ascii=False))
