--- analysisapi/scheduleplanner/web/property/config.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<config>
-  <diary_get_url>http://localhost:8080/api/diary</diary_get_url>
-  <diaries_get_url>http://localhost:8080/api/diary/list</diaries_get_url>    
-</config>
