+++ analysisapi/settings.py
@@ -1,0 +1,10 @@
+"""Analysis API の設定ファイルです。"""
+
+HOST = "127.0.0.1"
+PORT = 8000
+API_SPECIFICATION_FILE_PATH = "./api-docs/analysisapi/api-specification.json"
+PROJECT_NAME = "Analysis API"
+PROJECT_DESCRIPTION = "Analysis API の説明です。"
+PROJECT_VERSION = "1.0.0"
+DIARY_GET_URL = "http://localhost:8080/api/diary"
+DIARIES_GET_URL = "http://localhost:8080/api/diaries"
